breaking(command): remove `BaseCommand` class (#27)
All commands have to be created with the `Command` class for now.
The `help` and `completions` commands are now optional and can be registered as descripted in the example below.
The `--help` and `--version` option will be registered only on the main command for now. The `--help` option is a global option and available on all child-command's.

```typescript
import { Command, HelpCommand, CompletionsCommand } from 'https://deno.land/x/cliffy/command.ts';

await new Command()
    .command( 'help', new HelpCommand() )
    .command( 'completions', new CompletionsCommand() )
    .parse()
```