Resolve #223
Two parts to this fix. First if using Windows, the path begins with a DriveLetter: and not a forward slash. On Windows remove the first forward slash when creating paths for the file array. Second in the loop testing for commands, the loop returned false if the first command wasn't found, so it never searched the entire array. This will now find the file. I anticipate that if these commands are compiled the current method will have problems. It will likely work for Linux and MacOS because it is looking for a file name without a .ts extension, but it seems like it will also still run them as deno run instead of their compiled commands. TBD and I'll cross that bridge when I get there.