breaking(command): infer types and option names (#312)
With this release the `Command` class is now strict by default. All names and types from option, argument and environment variable definitions will be now automatically inferred 🚀

Previously it was necessary to pass the `void` type as first generic argument to the command class `Command<void>` (which is now the default) to enable strict typings.
The name and the type could than be specified by passing them to the `option`, `arguments` or `env` methods:

```ts
new Command<void>()
  .option<{ foo?: string }>("--foo <value:string>")
  .arguments<[string, string?]>("<input> [output]")
```

This is no longer necessary. Names and types are now automatically inferred from the definitions `"--foo <value:string>"` and you can define it just like the following example. Options and arguments are will be automattically typed.

```ts
new Command()
  .option("--foo <value:string>")
  .arguments("<input> [output]")
```

Dotted options, custom types like `EnumType` and options like `default`, `value` and `prefix` are supported as well.

To disable strict types you can pass the `any` type to the command class `Command<any>` which was previously the default.