feat(command): add new `secret` type  and `defaultText` option to override the display text for the default option value (#775)
This PR adds the following two features related to changing default
value display for options.

1. `defaultText` option for options

It is either a string or a function that transforms the default value
into string. When set, it changes the help text of the default, while
the underlying default value is used for processing.

2. `"secret"` input type

This is exactly the same as string type, except it hides its default
values on help text.

Resolves: #774